Initial commit of app

+++ app/src/main/java/com/scttech/android/kotlin/openfitness/data/local/entity/ProgramTestEntity.kt
@@ -30,7 +30,7 @@ data class ProgramTestEntity(
     val programId: Long,
     val profileId: Long,
     val testedAtEpochMillis: Long,
-    val result: Double,
+    val result: Int,
     val notes: String,
 )
